refactor highlight
