+++ web/app/components/Highlight.tsx
@@ -1,25 +1,10 @@
 import clsx from 'clsx'
-import { ReactNode } from 'react'
 
-export const Highlight = ({
-	children,
-	className,
-}: {
-	children: ReactNode | string
-	className?: string
-}) => {
-	return (
-		<>
-			{' '}
-			<span className={clsx('relative')}>
-				<div
-					className={clsx(
-						'absolute -inset-x-8 inset-y-0 -z-10 transform-gpu rounded-full blur-xl',
-						className,
-					)}
-				/>
-				{children}
-			</span>{' '}
-		</>
-	)
-}
+export const Highlight = ({ className }: { className?: string }) => (
+	<div
+		className={clsx(
+			'absolute -inset-x-8 inset-y-0 -z-10 transform-gpu rounded-full blur-xl will-change-[opacity]',
+			className,
+		)}
+	/>
+)
